% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileCount2.R
\name{computeLibSizeChrom}
\alias{computeLibSizeChrom}
\title{Perform overlap queries between reads and genome by sliding windows
Count reads over sliding windows.}
\usage{
computeLibSizeChrom(aln_list)
}
\arguments{
\item{aln_list}{a list.}
}
\value{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment} object with chromosome-level depth
The assays slot holds the counts, rowRanges holds the annotation from the
sliding widows of genome.
metadata contains lib.size.chrom for holding chromosome-level sequence depth
}
\description{
Perform overlap queries between reads and genome by sliding windows
Count reads over sliding windows.
}
\author{
Jun Yu,Hervé Pagès and Julie Zhu
}
