% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileCount.R
\name{countByOverlaps}
\alias{countByOverlaps}
\title{Count overlapping genomic ranges}
\usage{
countByOverlaps(features, reads, ignore.strand, inter.feature)
}
\arguments{
\item{features}{A object of \link[GenomicRanges]{GRanges} represents the 
feature regions to be counted.}

\item{reads}{object that represents the data to be counted. See 
\link[GenomicAlignments]{summarizeOverlaps}.}

\item{ignore.strand}{logical(1). ignore strand?}

\item{inter.feature}{not used. This parameter is required by 
\link[GenomicAlignments]{summarizeOverlaps}.}
}
\value{
return a vector of counts the same length as features.
}
\description{
Count the reads in a given feature. This function does not work for parallel.
}
