% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSignals.R
\name{exportSignals}
\alias{exportSignals}
\title{output signals to file}
\usage{
exportSignals(dat, assayName, colName, con, format = "bedGraph", ...)
}
\arguments{
\item{dat}{An object of \link[GenomicRanges]{GRanges}, 
or \link[SummarizedExperiment]{RangedSummarizedExperiment}
with assays of raw counts, ratios, background correct ratios,
smoothed ratios and z-scores. It should be an element of output of 
\link{smoothRatiosByChromosome}}

\item{assayName}{character(1). Assay name for 
\link[SummarizedExperiment]{RangedSummarizedExperiment}}

\item{colName}{character(1). Column name of metadata of dat or assay of dat 
for coverage weight, see
\link[GenomicRanges]{coverage}, 
\link[SummarizedExperiment]{RangedSummarizedExperiment}.}

\item{con}{The connection to which data is saved. If this is a character
vector, it is assumed to be a filename and a corresponding file connection
is created and then closed after exporting the object.
If missing, a \link[IRanges]{SimpleRleList} will be returned.}

\item{format}{The format of the output. see \link[rtracklayer]{export}.}

\item{...}{Parameters to pass to \link[rtracklayer]{export}}
}
\value{
If con is missing, a \link[IRanges]{SimpleRleList} will be returned.
Otherwise, nothing is returned.
}
\description{
Output signals to bedgraph, bed, wig, etc, for track viewer
}
\examples{
gr <- GRanges("chr1", IRanges(seq_len(100), 201:300), reads=rep(1, 100))
myTrackLine <- new("TrackLine", name="my track",
                    description="description of my track",
                    color=col2rgb("red")[, 1],
                    visibility="full")
exportSignals(gr, colName="reads", 
              con="test.bedGraph", trackLine=myTrackLine)
data(triplicates.counts)
exportSignals(triplicates.counts, "counts", 
              "G26.bam", "test.bw", format="bigWig")

}
