% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGenomeSAF.R
\name{generateGenomeSAF}
\alias{generateGenomeSAF}
\title{Generate sliding windows of a specified width and frequency}
\usage{
generateGenomeSAF(
  bamfile = NULL,
  windowSize = 50000L,
  step = 5000L,
  chrExcluded = c("chrM", "MT")
)
}
\arguments{
\item{bamfile}{A character(1), specifying a path to a BAM file, from which
chromosome names and lengths are extracted for generating sliding windows.}

\item{windowSize}{An integer(1), specifying the width of window in bp.}

\item{step}{An integer(1), specifying the step in bp for generating sliding
windows.}

\item{chrExcluded}{A character(n), specifying chromosomes to be excluded
from generating sliding windows.}
}
\value{
A data frame containing sliding windows for each chromosome and
individual chromosome in the SAF (simplified annotation format) format
as described in the Rsubread package.
}
\description{
Generate sliding windows of a specified width and frequency
}
\examples{
bamfiles <- list.files(system.file("extdata", package="NADfinder"),
                      recursive=FALSE,
                      pattern="*bam$",
                      full=TRUE)
genomeSW_SAF <-
    generateGenomeSAF(bamfile = bamfiles[1],
                      windowSize = 50000L,
                      step = 5000L,
                      chrExcluded = c("chrM", "MT"))
}
