% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2ratio.R
\name{log2ratio}
\alias{log2ratio}
\title{calculate the log2 transformed ratios}
\usage{
log2ratio(A, B, pseudocount)
}
\arguments{
\item{A, B}{counts for nucleosome and genome. They should be
numeric vectors with identical length.}

\item{pseudocount}{pseudo-count will be used to aviod x/0 by x/pseudocount.
If it is not set, pseudo-count will be the minimal count except 0 of inputs.}
}
\value{
A vector of numeric of log2 transformed ratios.
}
\description{
calculate the log2 transformed ratios for nucleosome vs genome.
pseudo-count will be used to avoid x/0.
}
\examples{
log2ratio(seq_len(10), 10:1)

}
