% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.count.R
\docType{data}
\name{single.count}
\alias{single.count}
\title{Counts data for chromosome 18 for an experiment of a single pair of samples}
\description{
Counts data for chromosome 18 for an experiment of a single pair of samples
}
\keyword{data}
