% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplicates.counts.R
\docType{data}
\name{triplicate.count}
\alias{triplicate.count}
\title{Counts data for chromosome 18 for an expriment with triplicates}
\description{
Counts data for chromosome 18 for an expriment with triplicates
}
\keyword{data}
