% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreOverBck.R
\name{zscoreOverBck}
\alias{zscoreOverBck}
\title{Z-scores over the background}
\usage{
zscoreOverBck(ratios, backgroundPercentage = 0.25)
}
\arguments{
\item{ratios}{A numeric vector containing the transformed, background corrected and smoothed ratios in each window.}

\item{backgroundPercentage}{numeric(1). Low percentile for background distribution.}
}
\value{
A vector of numeric. Z-scores.
}
\description{
Calculate the z-scores over the background distribution.
}
\examples{
r <- runif(200)
zscoreOverBck(r)
}
\author{
Jianhong Ou and Julie Zhu
}
