% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{Call peaks using transformed, background corrected, and smoothed ratios with biological replicates}
\usage{
callPeaks(se, backgroundCorrectedAssay = "bcRatio",
  normlizationMethod = "quantile", N = 100, cutoffAdjPvalue = 1e-04,
  countFilter = 1000, method.combineP = "meansig", ...)
}
\arguments{
\item{se}{An object of 
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment}
with assays of raw counts, tranformed ratios, background corrected ratios,
smoothed ratios and z-scores. It should be an element of output of 
\link{smoothRatiosByChromosome}}

\item{backgroundCorrectedAssay}{character(1). Assays names 
for background corrected log2-transformed ratios, CPMRatios or OddRatios.}

\item{normlizationMethod}{character(1) specifying the normalization
method to be used. Choices  are "none", "scale", "quantile" or "cyclicloess".
See \link[limma]{normalizeBetweenArrays} for details.}

\item{N}{numeric(1) or integer(1).
Critical frequencies of the low pass filter will be set as 1/N.
1/N is a cutoff at 1/N-th of the Nyquist frequency.
Default 100.}

\item{cutoffAdjPvalue}{numeric(1). Cutoff adjust p-value.}

\item{countFilter}{numeric(1). Cutoff value for mean of raw reads count 
in each window.}

\item{method.combineP}{A method used to combine P-values. Default meansig}

\item{...}{Parameter not used.}
}
\value{
An object of GRanges of peak list with metadata "AveSig", "P.Value", 
and "adj.P.Val", where "AveSig" means average signal such as average log2OddsRatio, log2CPMRatio or log2Ratio.
}
\description{
Use limma to calculate p-values for NADs
}
\details{
By default, use the mean smoothed ratio for each peak region to calculate p-values
}
\examples{

data(triplicates.counts)
se <- triplicates.counts
gps <- c("26", "28", "29")
se <- log2se(se,transformation = "log2Ratio", 
             nucleoleusCols = paste0("N", gps, ".bam"),
             genomeCols = paste0("G", gps, ".bam"))
se<- smoothRatiosByChromosome(se, chr="chr18")
peaks <- callPeaks(se[[1]][10000:15000, ], 
                cutoffAdjPvalue=0.001, countFilter=1000)

}
\author{
Jianhong Ou, Haibo Liu and Julie Zhu
}
