% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2se.R
\name{log2se}
\alias{log2se}
\title{calculate the log2 transformed ratios for SummarizedExperiment class}
\usage{
log2se(se, nucleosomeCols, genomeCols, pseudocount)
}
\arguments{
\item{se}{A \link[SummarizedExperiment]{RangedSummarizedExperiment} object.
The output of \link{tileCount}.}

\item{nucleosomeCols, genomeCols}{column Names of counts for nucleosome 
and genome. They should be the column names in the assays of se. 
Ratios will be calculated as log2(nucleosomeCols/genomeCols).}

\item{pseudocount}{pseudo-count will be used to aviod x/0 by x/pseudocount.
If it is not set, pseudo-count will be the minimal count except 0 of inputs.}
}
\value{
A RangedSummarizedExperiment object with log2 transformed ratios. 
Assays will be named as nucleosome, genome and ratio.
}
\description{
Calculate the log2 transformed ratios for nucleosome vs genome.
pseudo-count will be used to avoid x/0.
}
\examples{
library(SummarizedExperiment)
se <- 
 SummarizedExperiment(assays=list(counts=DataFrame(A=seq_len(3), 
                                                   B=rep(1, 3))), 
                      rowRanges=GRanges("chr1", 
                                         IRanges(c(1, 10, 20), 
                                                 width=9)))
log2se(se, "A", "B")

}
