% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NADfinder.R
\docType{package}
\name{NADfinder-package}
\alias{NADfinder}
\alias{NADfinder-package}
\title{Identify nucleolus-associated domains (NADs) from NAD-seq}
\description{
Sliding-window based peak calling algorithm using whole genome sequences as control.
}
