% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreOverBck.R
\name{zscoreOverBck}
\alias{zscoreOverBck}
\title{Z-scores over the background}
\usage{
zscoreOverBck(ratios, backgroundPercentage = 0.25)
}
\arguments{
\item{ratios}{A vector of numeric. It is the ratios of counts in each window.}

\item{backgroundPercentage}{numeric(1). Percentage of value for background.}
}
\value{
A vector of numeric. Z-scores.
}
\description{
Calculate the z-scores over the lower percentage values.
}
\examples{
r <- runif(200)
zscoreOverBck(r)

}
